feat: add finalizer for handling the deletion (#20)
When a resource is deleted, the tag binding k8s resources will be
deleted after the resource (due to how ownership and async cleanup works
in K8s). Deleting the tag bindings then fails, because they don't exist
anymore (resource was deleted). This leads to stuck tag binding
resources in the cluster. We should make sure that resources can get
cleanly deleted, including their generated tag bindings.

We should check if the resource is already being deleted from GCP when a
finalizer is still on there (it probably is). We orphan the tag binding
resources when the resource is deleted.

---------

Co-authored-by: Heiko Rothe <heiko.rothe@deliveryhero.com>