fix: Cleanup tags when component is removed (#28)
When we receive a deletion request for the resource, we will also try to
clean the TagBindings associated, which helps clean up unused
tagValues/tagkeys residing on the GCP project

Tested locally the following scenarios:
----------------------------------------
Multiple Resources created with multiple tags and deleting the resources
only deletes the unused tags leaving behind the used tags(Since it is
binding to other resources)

Tested change in tag key-values which creates the new ones and deletes
the old ones if unused.

Creation and deletion of tags for single resources.

---------

Co-authored-by: jijinp <jijin.puthuvakkattil@deliveryhero.com>